refactor(challenger): simplify test utilities
